I've added a new "URL Downloader" feature to the app, which you can find in the "ฟีเจอร์เสริม & มินิเกม" (Extras & Minigames) section.
This new tool allows you to paste a direct URL to a video, image, or audio file. The app will then attempt to display a preview of the media and provide a button to download it. I noticed you provided a YouTube URL; while this tool won't be able to process YouTube links directly due to their terms of service and technical restrictions, it will work perfectly for direct links to media files you find across the web.

Here are the changes I've made: